Add the code from one of my dead projects
I was doing a program that computes hashes using different algorithms,
but I was a little bit drained after implementing md5 and abandoned it.

I though it would be a good idea to isolate my md5 implementaion and
release it as a library crate... and here I am.

I needs a lot of cleanup and optimization, but the base implementation
works.